+++ src/PluginManager/include/PluginManager/Connect/AbstractSkyConnect.h
@@ -126,7 +126,20 @@ protected:
     virtual bool onFreezeUserAircraft(bool enable) const noexcept = 0;
     virtual bool onSimulationEvent(SimulationEvent event, float arg1) const noexcept = 0;
 
-    virtual bool onStartRecording() noexcept = 0;
+    /*!
+     * Called when recording a Flight has started.
+     *
+     * \return \c true on success; \c false on error (SimConnect connection error)
+     */
+    virtual bool onStartFlightRecording() noexcept = 0;
+
+    /*!
+     * Called when recording an Aircraft has started, excluding Flight data.
+     * This method is called when recording a formation aircraft.
+     *
+     * \return \c true on success; \c false on error (SimConnect connection error)
+     */
+    virtual bool onStartAircraftRecording() noexcept = 0;
     virtual void onRecordingPaused(bool paused) noexcept = 0;
     virtual void onStopRecording() noexcept = 0;