+++ src/Plugins/Connect/PathCreator/src/PathCreatorPlugin.h
@@ -62,7 +62,8 @@ protected:
     bool onFreezeUserAircraft(bool enable) const noexcept override;
     bool onSimulationEvent(SimulationEvent event, float arg1) const noexcept override;
 
-    bool onStartRecording() noexcept override;
+    bool onStartFlightRecording() noexcept override;
+    bool onStartAircraftRecording() noexcept override;
     void onRecordingPaused(bool paused) noexcept override;
     void onStopRecording() noexcept override;